Using the Keras Library on the MNIST Dataset
Objective:

The objective of this project is to leverage the Keras library to develop, train, and evaluate a machine learning model on the MNIST dataset. The MNIST dataset is a well-known benchmark in the field of machine learning and computer vision, consisting of 70,000 images of handwritten digits (0-9). The goal is to achieve high accuracy in classifying these digits.

Tools and Libraries:

Keras: A powerful and user-friendly deep learning library that runs on top of TensorFlow.
MNIST Dataset: A large database of handwritten digits used for training and testing various machine learning models.
Python: The programming language used for developing the project.